+++ fabric/contrib/project.py
@@ -17,7 +17,7 @@ __all__ = ['rsync_project', 'upload_project']
 
 @needs_host
 def rsync_project(remote_dir, local_dir=None, exclude=(), delete=False,
-    extra_opts='', ssh_opts=''):
+    extra_opts='', ssh_opts='', capture=False):
     """
     Synchronize a remote directory with the current project directory via rsync.
 
@@ -67,6 +67,7 @@ def rsync_project(remote_dir, local_dir=None, exclude=(), delete=False,
       custom arguments or options to ``rsync``.
     * ``ssh_opts``: Like ``extra_opts`` but specifically for the SSH options
       string (rsync's ``--rsh`` flag.)
+    * ``capture``: Sent directly into an inner `~fabric.operations.local` call.
 
     Furthermore, this function transparently honors Fabric's port and SSH key
     settings. Calling this function when the current host string contains a
@@ -79,8 +80,10 @@ def rsync_project(remote_dir, local_dir=None, exclude=(), delete=False,
         rsync [--delete] [--exclude exclude[0][, --exclude[1][, ...]]] \\
             -pthrvz [extra_opts] <local_dir> <host_string>:<remote_dir>
 
-    .. versionadded:: 1.4
+    .. versionadded:: 1.4.0
         The ``ssh_opts`` keyword argument.
+    .. versionadded:: 1.4.1
+        The ``capture`` keyword argument.
     """
     # Turn single-string exclude into a one-item list for consistency
     if not hasattr(exclude, '__iter__'):
@@ -117,7 +120,7 @@ def rsync_project(remote_dir, local_dir=None, exclude=(), delete=False,
     cmd = "rsync %s %s %s@%s:%s" % (options, local_dir, user, host, remote_dir)
     if output.running:
         print("[%s] rsync_project: %s" % (env.host_string, cmd))
-    return local(cmd)
+    return local(cmd, capture=capture)
 
 
 def upload_project(local_dir=None, remote_dir=""):